Remove nvm from the Powerlevel10k element list
This and the node_version element display the same information. Only the
display condition differs. The node_version element shows this
information only if the current directory is part of a Node.js project.
This is the desired behavior. Therefore there is no need to also enable
the nvm element.