Integrate Microsoft Teams PWA
This commits adds a *.desktop file to start the Microsoft Teams PWA in
Chromiums app mode. It also starts Chromium with the Wayland backend
enabled if it is possible to use it and PipeWire for screen capture. The
Chromium app mode inhibits all shortcuts. Therefore the shortcut
inhibition is also disabled for the Microsoft Teams PWA to be able to
use Sway normally.